DEV -> MAIN huge update

+++ src/modules/petition-users/petition-users.service.ts
@@ -16,7 +16,7 @@ export const loginUser = async (data: LoginDTO) => {
   const token = generateToken({
     username: user.username,
     userId: user._id.toString(),
-    role: "petition_user",
+    role: user.role,
   });
 
   return token;
@@ -36,6 +36,7 @@ export const registerUser = async (data: RegisterDTO) => {
     passwordHash,
     name,
     surname,
+    role: "petition_user" //nie musi tego byc bo w modelu jest default ale dodaje tutaj
   });
 
   //Tutaj jest generowane po ._id ale mozna tez po username
@@ -48,7 +49,10 @@ export const registerUser = async (data: RegisterDTO) => {
   return token;
 };
 
-export const deleteUserService = async (userId: string) => {
+export const deleteUserService = async (userId: string, userRole: string) => {
+  if (userRole !== "petition_user") {
+    throw new Error("Brak uprawnień: tylko zwykli użytkownicy mogą usunąć swoje konto");
+  }
   const deleted = await PetitionUserModel.findByIdAndDelete(userId);
   if (!deleted) throw new Error("Użytkownik nie istnieje");
